test(refactor): Refactor point representation tests
- Introduced AbstractPointRepresentationTest to consolidate tests for point types.
- Removed the old AbstractPointTest which was overly complex and redundant.
- Created AbstractPointTestCase for shared test functionality among point tests.
- Updated tests to ensure proper coverage for JSON serialization, SRID handling, and coordinate conversions.
