--- tests/LongitudeOne/Spatial/Tests/PHP/Types/AbstractPointTest.php
@@ -1,736 +1,0 @@
-<?php
-/**
- * This file is part of the Doctrine Spatial extension.
- *
- * PHP 8.4 | 8.5
- * Doctrine ORM ^3.6
- *
- * Copyright Alexandre Tranchant <alexandre.tranchant@gmail.com> 2017-2026
- * Copyright Longitude One 2020-2026
- * Copyright 2015 Derek J. Lambert
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- *
- */
-
-declare(strict_types=1);
-
-namespace LongitudeOne\Spatial\Tests\PHP\Types;
-
-use LongitudeOne\Spatial\Exception\InvalidValueException;
-use LongitudeOne\Spatial\PHP\Types\AbstractPoint;
-use LongitudeOne\Spatial\PHP\Types\Geography\Point as GeographicPoint;
-use LongitudeOne\Spatial\PHP\Types\Geometry\Point as GeometricPoint;
-use LongitudeOne\Spatial\Tests\DataProvider as LoDataProvider;
-use LongitudeOne\Spatial\Tests\Helper\PointHelperTrait;
-use PHPUnit\Framework\Attributes\DataProvider;
-use PHPUnit\Framework\Attributes\Group;
-use PHPUnit\Framework\TestCase;
-
-/**
- * Geometric and geographic points tests.
- * These methods involve tests launched on both Geometric and Geographic points.
- *
- * @internal
- *
- * @covers \LongitudeOne\Spatial\PHP\Types\AbstractPoint
- * @covers \LongitudeOne\Spatial\PHP\Types\Geography\Point
- * @covers \LongitudeOne\Spatial\PHP\Types\Geometry\Point
- */
-#[Group('php')]
-class AbstractPointTest extends TestCase
-{
-    // phpcs:disable Squiz.Commenting.FunctionComment.IncorrectTypeHint
-
-    use PointHelperTrait;
-
-    private const DELTA = 0.000000000001;
-
-    /**
-     * Assert that the object has the method.
-     *
-     * @param object $object object to test
-     * @param string $method the method to test
-     */
-    private static function assertObjectHasMethod(object $object, string $method): void
-    {
-        static::assertTrue(method_exists($object, $method), sprintf('Method "%s":"%s" does not exist.', $object::class, $method));
-    }
-
-    /**
-     * Test bad array parameter - Object as value.
-     *
-     * @param class-string<AbstractPoint> $abstractPoint Geometric point and geographic point
-     */
-    #[DataProvider('pointTypeProvider')]
-    public function testArrayWithObject(string $abstractPoint): void
-    {
-        $this->expectException(InvalidValueException::class);
-        $this->expectExceptionMessage(sprintf('Invalid parameters passed to %s::__construct: array(object, NULL)', $abstractPoint));
-
-        new $abstractPoint([new \stdClass(), null]);
-    }
-
-    /**
-     * SRID values passed to constructors must be integers or null.
-     *
-     * @param class-string<AbstractPoint> $pointType Geometric point and geographic point
-     * @param mixed                       $srid      invalid SRID value
-     */
-    #[DataProvider('invalidSridProvider')]
-    public function testConstructorRejectsNonIntegerSrid(string $pointType, mixed $srid): void
-    {
-        self::expectException(InvalidValueException::class);
-
-        new $pointType('1', '2', $srid);
-    }
-
-    /**
-     * @return \Generator<string, array{0: class-string<AbstractPoint>, 1: mixed}, null, void>
-     */
-    public static function invalidSridProvider(): \Generator
-    {
-        $invalidSrids = [
-            'numeric string' => '4326',
-            'float' => 4326.9,
-        ];
-
-        foreach (self::pointTypeProvider() as $pointName => [$pointType]) {
-            foreach ($invalidSrids as $sridName => $srid) {
-                yield sprintf('%s with %s SRID', $pointName, $sridName) => [$pointType, $srid];
-            }
-        }
-    }
-
-    /**
-     * Test getType method.
-     *
-     * @param class-string<AbstractPoint> $class the classname to test, Geometric point and geographic point
-     */
-    #[DataProvider('pointTypeProvider')]
-    public function testGetType(string $class): void
-    {
-        $point = new $class(0, 0);
-        static::assertEquals('Point', $point->getType());
-    }
-
-    /**
-     * Test geodesic setters.
-     *
-     * @param class-string<AbstractPoint> $pointType         Geometric or geographic point
-     * @param float|int|string            $longitude         the actual longitude
-     * @param float|int|string            $latitude          the actual latitude
-     * @param float|int                   $expectedLongitude the expected longitude
-     * @param float|int                   $expectedLatitude  the expected latitude
-     *
-     * @throws InvalidValueException It shall NOT happen in this test
-     */
-    #[DataProvider('validGeodesicCoordinateProvider')]
-    public function testGoodGeodesicCoordinate(string $pointType, float|int|string $longitude, float|int|string $latitude, float|int $expectedLongitude, float|int $expectedLatitude): void
-    {
-        $geographicPoint = new $pointType(0, 0);
-        $geographicPoint->setLongitude($longitude);
-        $geographicPoint->setLatitude($latitude);
-
-        static::assertSame($expectedLongitude, $geographicPoint->getLongitude());
-        static::assertSame($expectedLatitude, $geographicPoint->getLatitude());
-    }
-
-    /**
-     * @return \Generator<string, array{0: class-string<AbstractPoint>, 1: float|int|string, 2: float|int|string, 3: float|int, 4: float|int}, null, void>
-     */
-    public static function validGeodesicCoordinateProvider(): \Generator
-    {
-        $points = [
-            'GeometricPoint' => GeometricPoint::class,
-            'GeographicPoint' => GeographicPoint::class,
-        ];
-        foreach ($points as $point) {
-            foreach (LoDataProvider::validGeodesicCoordinateProvider() as $key => $value) {
-                yield sprintf('%s(%s)', $point, $key) => array_merge([$point], $value);
-            }
-        }
-    }
-
-    /**
-     * Test valid string points.
-     *
-     * @param class-string<AbstractPoint> $abstractPoint     Geometric point and geographic point
-     * @param float|int|string            $longitude         the longitude to test
-     * @param float|int|string            $latitude          the latitude to test
-     * @param float|int                   $expectedLongitude the expected longitude
-     * @param float|int                   $expectedLatitude  the expected latitude
-     */
-    #[DataProvider('mixedGeodesicCoordinateProvider')]
-    public function testGoodStringPoints(string $abstractPoint, float|int|string $longitude, float|int|string $latitude, float|int $expectedLongitude, float|int $expectedLatitude): void
-    {
-        $point = new $abstractPoint($longitude, $latitude);
-        static::assertEqualsWithDelta($expectedLongitude, $point->getLongitude(), self::DELTA);
-        static::assertEqualsWithDelta($expectedLatitude, $point->getLatitude(), self::DELTA);
-
-        $point = new $abstractPoint(0, 0, 4326);
-        $point->setLongitude($longitude);
-        $point->setLatitude($latitude);
-        static::assertEqualsWithDelta($expectedLongitude, $point->getLongitude(), self::DELTA);
-        static::assertEqualsWithDelta($expectedLatitude, $point->getLatitude(), self::DELTA);
-    }
-
-    /**
-     * @return \Generator<string, array{0: class-string, 1: float|int|string, 2: float|int|string, 3: float|int, 4: float|int}, null, void>
-     */
-    public static function mixedGeodesicCoordinateProvider(): \Generator
-    {
-        $points = [
-            'GeometricPoint' => GeometricPoint::class,
-            'GeographicPoint' => GeographicPoint::class,
-        ];
-
-        /** @var array<string, array{0: float, 1: float}> $geodesicCoordinates */
-        $geodesicCoordinates = [
-            '79:56:55W, 40:26:46N' => [-79.9486111111111, 40.44611111111111],
-            '79° 56\' 55" W, 40° 26\' 46" N' => [-79.9486111111111, 40.44611111111111],
-            '79°56′55″W, 40°26′46″N' => [-79.9486111111111, 40.44611111111111],
-            '79° 56′ 55″ W, 40° 26′ 46″ N' => [-79.9486111111111, 40.44611111111111],
-            '79:56:55.832W, 40:26:46.543N' => [-79.94884222222223, 40.446261944444444],
-            '112:4:0W, 33:27:0N' => [-112.066666666666, 33.45],
-        ];
-
-        foreach ($points as $className => $class) {
-            foreach ($geodesicCoordinates as $coordinatesString => $expected) {
-                $coordinates = explode(', ', $coordinatesString);
-
-                yield sprintf('%s(%s, %s)', $className, $coordinates[0], $coordinates[1]) => [$class, $coordinates[0], $coordinates[1], $expected[0], $expected[1]];
-            }
-        }
-    }
-
-    /**
-     * Test argument 1 with too few values - Two invalid parameters.
-     *
-     * @param class-string<AbstractPoint> $abstractPoint Geometric point and geographic point
-     */
-    #[DataProvider('pointTypeProvider')]
-    public function testInvalidArrayWithTooFewValues(string $abstractPoint): void
-    {
-        $this->expectException(InvalidValueException::class);
-        $this->expectExceptionMessage(sprintf('Invalid parameters passed to %s::__construct: array(1)', $abstractPoint));
-
-        new $abstractPoint([1]);
-    }
-
-    /**
-     * Test argument 1 with too many values - Two invalid parameters.
-     *
-     * @param class-string<AbstractPoint> $abstractPoint Geometric point and geographic point
-     */
-    #[DataProvider('pointTypeProvider')]
-    public function testInvalidArrayWithTooManyValues(string $abstractPoint): void
-    {
-        $this->expectException(InvalidValueException::class);
-        $this->expectExceptionMessage(sprintf('Invalid parameters passed to %s::__construct: array(1, 2, 3, 4)', $abstractPoint));
-
-        new $abstractPoint([1, 2, 3, 4]);
-    }
-
-    /**
-     * Coordinate values in exceptions must be bounded and free of control characters.
-     *
-     * @param class-string<AbstractPoint> $pointType  Geometric point and geographic point
-     * @param string                      $coordinate invalid coordinate value
-     */
-    #[DataProvider('unsafeCoordinateProvider')]
-    public function testInvalidCoordinateExceptionDoesNotExposeUnsafeInput(string $pointType, string $coordinate): void
-    {
-        try {
-            new $pointType($coordinate, '2');
-            static::fail('An invalid coordinate must throw an exception.');
-        } catch (InvalidValueException $exception) {
-            static::assertLessThan(1024, mb_strlen($exception->getMessage()));
-            static::assertStringNotContainsString("\r", $exception->getMessage());
-            static::assertStringNotContainsString("\n", $exception->getMessage());
-        }
-    }
-
-    /**
-     * @return \Generator<string, array{0: class-string<AbstractPoint>, 1: string}, null, void>
-     */
-    public static function unsafeCoordinateProvider(): \Generator
-    {
-        $coordinates = [
-            'long value' => str_repeat('invalid-coordinate-', 1000),
-            'control characters' => "invalid\r\ncoordinate",
-        ];
-
-        foreach (self::pointTypeProvider() as $pointName => [$pointType]) {
-            foreach ($coordinates as $coordinateName => $coordinate) {
-                yield sprintf('%s with %s', $pointName, $coordinateName) => [$pointType, $coordinate];
-            }
-        }
-    }
-
-    /**
-     * Test to convert point to json.
-     *
-     * @param class-string<AbstractPoint> $abstractPoint Geometric point and geographic point
-     */
-    #[DataProvider('pointTypeProvider')]
-    public function testJson(string $abstractPoint): void
-    {
-        $expected = '{"type":"Point","coordinates":[5,5],"srid":null}';
-        $point = new $abstractPoint(5, 5);
-
-        static::assertEquals($expected, $point->toJson());
-        static::assertEquals($expected, json_encode($point));
-
-        $point->setSrid(4326);
-        $expected = '{"type":"Point","coordinates":[5,5],"srid":4326}';
-        static::assertEquals($expected, $point->toJson());
-        static::assertEquals($expected, json_encode($point));
-    }
-
-    /**
-     * Deprecated array-based constructors must reject associative arrays rather than passing named arguments.
-     *
-     * @param class-string<AbstractPoint> $pointType Geometric point and geographic point
-     */
-    #[DataProvider('pointTypeProvider')]
-    public function testLegacyArrayWithAssociativeKeys(string $pointType): void
-    {
-        self::expectException(InvalidValueException::class);
-
-        new $pointType(['unexpected' => '1', 'other' => '2']);
-    }
-
-    /**
-     * Deprecated array-based constructors must validate every coordinate before invoking construct().
-     *
-     * @param class-string<AbstractPoint> $pointType         Geometric point and geographic point
-     * @param mixed                       $invalidCoordinate invalid coordinate value
-     */
-    #[DataProvider('invalidLegacyArrayCoordinateProvider')]
-    public function testLegacyArrayWithInvalidCoordinate(string $pointType, mixed $invalidCoordinate): void
-    {
-        self::expectException(InvalidValueException::class);
-
-        new $pointType([$invalidCoordinate, '2'], 4326);
-    }
-
-    /**
-     * @return \Generator<string, array{0: class-string<AbstractPoint>, 1: mixed}, null, void>
-     */
-    public static function invalidLegacyArrayCoordinateProvider(): \Generator
-    {
-        $invalidCoordinates = [
-            'nested array' => ['1'],
-            'object' => new \stdClass(),
-            'closure' => static function (): void {
-            },
-            'boolean' => true,
-        ];
-
-        foreach (self::pointTypeProvider() as $pointName => [$pointType]) {
-            foreach ($invalidCoordinates as $coordinateName => $invalidCoordinate) {
-                yield sprintf('%s with %s', $pointName, $coordinateName) => [$pointType, $invalidCoordinate];
-            }
-        }
-    }
-
-    /**
-     * Deprecated array-based constructors must reject resources before invoking construct().
-     *
-     * @param class-string<AbstractPoint> $pointType Geometric point and geographic point
-     */
-    #[DataProvider('pointTypeProvider')]
-    public function testLegacyArrayWithResource(string $pointType): void
-    {
-        $resource = fopen('php://memory', 'r');
-        if (false === $resource) {
-            static::fail('Unable to open the in-memory resource.');
-        }
-
-        try {
-            self::expectException(InvalidValueException::class);
-
-            new $pointType([$resource, '2'], 4326);
-        } finally {
-            fclose($resource);
-        }
-    }
-
-    /**
-     * Deprecated array-based constructors with a separate SRID must not reinterpret malformed arrays.
-     *
-     * @param class-string<AbstractPoint> $pointType   Geometric point and geographic point
-     * @param mixed[]                     $coordinates malformed coordinate array
-     */
-    #[DataProvider('invalidLegacyArrayWithSridProvider')]
-    public function testLegacyArrayWithSridHasExactlyTwoCoordinates(string $pointType, array $coordinates): void
-    {
-        self::expectException(InvalidValueException::class);
-
-        new $pointType($coordinates, 4326);
-    }
-
-    /**
-     * @return \Generator<string, array{0: class-string<AbstractPoint>, 1: mixed[]}, null, void>
-     */
-    public static function invalidLegacyArrayWithSridProvider(): \Generator
-    {
-        $invalidCoordinates = [
-            'no coordinate' => [],
-            'one coordinate' => ['1'],
-            'three coordinates' => ['1', '2', 3857],
-            'four coordinates' => ['1', '2', 3857, 9999],
-        ];
-
-        foreach (self::pointTypeProvider() as $pointName => [$pointType]) {
-            foreach ($invalidCoordinates as $coordinatesName => $coordinates) {
-                yield sprintf('%s with %s', $pointName, $coordinatesName) => [$pointType, $coordinates];
-            }
-        }
-    }
-
-    /**
-     * Test bad string parameters - No parameters.
-     *
-     * @param class-string<AbstractPoint> $pointType Geometric point and geographic point
-     */
-    #[DataProvider('pointTypeProvider')]
-    public function testMissingArguments(string $pointType): void
-    {
-        $this->expectException(InvalidValueException::class);
-        $this->expectExceptionMessage(sprintf('Invalid parameters passed to %s::__construct:', $pointType));
-
-        new $pointType();
-    }
-
-    /**
-     * Test a point created with an array.
-     *
-     * @param class-string<AbstractPoint> $pointType Geometric point and geographic point
-     */
-    #[DataProvider('pointTypeProvider')]
-    public function testPointFromArrayToString(string $pointType): void
-    {
-        $point = new $pointType([5, 5]);
-
-        static::assertSame('5 5', (string) $point);
-    }
-
-    /**
-     * Test error when point is created with too many arguments.
-     *
-     * @param class-string<AbstractPoint> $abstractPoint Geometric point and geographic point
-     */
-    #[DataProvider('pointTypeProvider')]
-    public function testPointTooManyArguments(string $abstractPoint): void
-    {
-        $this->expectException(InvalidValueException::class);
-        $this->expectExceptionMessage(sprintf('Invalid parameters passed to %s::__construct: 5, 5, 5, 5', $abstractPoint));
-
-        new $abstractPoint(5, 5, 5, 5);
-    }
-
-    /**
-     * Test point with srid.
-     *
-     * @param class-string<AbstractPoint> $abstractPoint Geometric point and geographic point
-     */
-    #[DataProvider('pointTypeProvider')]
-    public function testPointWithSrid(string $abstractPoint): void
-    {
-        $point = new $abstractPoint(5, 5, 2154);
-        $actual = $point->getSrid();
-        static::assertSame(2154, $actual);
-
-        $point->setSrid(4326);
-        $actual = $point->getSrid();
-        static::assertSame(4326, $actual);
-    }
-
-    /**
-     * Test error when point was created with the wrong arguments type.
-     *
-     * @param class-string<AbstractPoint> $abstractPoint Geometric point and geographic point
-     */
-    #[DataProvider('pointTypeProvider')]
-    public function testPointWrongArgumentTypes(string $abstractPoint): void
-    {
-        $this->expectException(InvalidValueException::class);
-        $this->expectExceptionMessage(sprintf('Invalid parameters passed to %s::__construct: array, array, 1234', $abstractPoint));
-
-        new $abstractPoint([], [], '1234');
-    }
-
-    /**
-     * This test checks that the geo-parser range exceptions are caught and "converted" to InvalidValueException.
-     * This test focuses on constructor.
-     *
-     * @param class-string<AbstractPoint> $abstractPoint    Geometric point and geographic point
-     * @param string                      $firstCoordinate  the first coordinate to test
-     * @param string                      $secondCoordinate the second coordinate to test
-     * @param string                      $expectedMessage  the expected message
-     */
-    #[DataProvider('rangeExceptionProvider')]
-    public function testRangeExceptionAreCaughtWithConstructor(string $abstractPoint, string $firstCoordinate, string $secondCoordinate, string $expectedMessage): void
-    {
-        $this->expectException(InvalidValueException::class);
-        $this->expectExceptionMessage($expectedMessage);
-
-        new $abstractPoint($firstCoordinate, $secondCoordinate);
-    }
-
-    /**
-     * This test checks that the geo-parser range exceptions are caught and "converted" to InvalidValueException.
-     * This test focuses on X and Y setters.
-     *
-     * @param class-string<AbstractPoint> $abstractPoint    Geometric point and geographic point
-     * @param string                      $firstCoordinate  the first coordinate to test
-     * @param string                      $secondCoordinate the second coordinate to test
-     * @param string                      $expectedMessage  the expected message
-     */
-    #[DataProvider('rangeExceptionProvider')]
-    public function testRangeExceptionAreCaughtWithNonExpectedSetters(string $abstractPoint, string $firstCoordinate, string $secondCoordinate, string $expectedMessage): void
-    {
-        $this->expectException(InvalidValueException::class);
-        $this->expectExceptionMessage($expectedMessage);
-
-        $point = new $abstractPoint(0, 0);
-        $point->setX($firstCoordinate);
-        $point->setY($secondCoordinate);
-    }
-
-    /**
-     * This test checks that the geo-parser range exceptions are caught and "converted" to InvalidValueException.
-     * This test focuses on longitude and latitude setters.
-     *
-     * @param class-string<AbstractPoint> $abstractPoint    Geometric point and geographic point
-     * @param string                      $firstCoordinate  the first coordinate to test
-     * @param string                      $secondCoordinate the second coordinate to test
-     * @param string                      $expectedMessage  the expected message
-     */
-    #[DataProvider('rangeExceptionProvider')]
-    public function testRangeExceptionAreCaughtWithSetters(string $abstractPoint, string $firstCoordinate, string $secondCoordinate, string $expectedMessage): void
-    {
-        $this->expectException(InvalidValueException::class);
-        $this->expectExceptionMessage($expectedMessage);
-
-        $point = new $abstractPoint(0, 0);
-        $point->setLongitude($firstCoordinate);
-        $point->setLatitude($secondCoordinate);
-    }
-
-    /**
-     * @return \Generator<string, array{0: class-string<AbstractPoint>, 1: string, 2: string, 3: string}, null, void>
-     */
-    public static function rangeExceptionProvider(): \Generator
-    {
-        $points = [
-            'GeometricPoint' => GeometricPoint::class,
-            'GeographicPoint' => GeographicPoint::class,
-        ];
-
-        $exceptions = [
-            'Out of range latitude' => [
-                '79:56:55W', '92:26:46N',
-                'Out of range latitude value, latitude must be between -90 and 90, got "92:26:46N".',
-            ],
-            'Out of range longitude' => [
-                '190:56:55W', '84:26:46N',
-                'Out of range longitude value, longitude must be between -180 and 180, got "190:56:55W".',
-            ],
-            'Invalid latitude direction' => [
-                '100:56:55W', '84:26:46Q',
-                'Invalid coordinate value, got "84:26:46Q".',
-            ],
-            'Latitude minutes greater than 59' => [
-                '108:42:55W', '84:64:46N',
-                'Out of range minute value, minute must be between 0 and 59, got "84:64:46N".',
-            ],
-            'Latitude seconds greater than 59' => [
-                '108:42:55W', '84:23:75N',
-                'Out of range second value, second must be between 0 and 59, got "84:23:75N".',
-            ],
-            'Longitude degrees greater than 180' => [
-                '190:56:55W', '84:26:46N',
-                'Out of range longitude value, longitude must be between -180 and 180, got "190:56:55W".',
-            ],
-            'Invalid longitude direction' => [
-                '100:56:55P', '84:26:46N',
-                'Invalid coordinate value, got "100:56:55P".',
-            ],
-            'Longitude minutes greater than 59' => [
-                '108:62:55W', '84:26:46N',
-                'Out of range minute value, minute must be between 0 and 59, got "108:62:55W".',
-            ],
-            'Longitude seconds greater than 59' => [
-                '108:53:94W', '84:26:46N',
-                'Out of range second value, second must be between 0 and 59, got "108:53:94W".',
-            ],
-        ];
-
-        foreach ($points as $className => $class) {
-            foreach ($exceptions as $dataTestName => $dataTest) {
-                yield sprintf('%s with a %s', $dataTestName, $className) => [$class, $dataTest[0], $dataTest[1], $dataTest[2]];
-            }
-        }
-    }
-
-    /**
-     * Test setX method.
-     *
-     * @param class-string<AbstractPoint> $abstractPoint Geometric point and geographic point
-     * @param string                      $setter        the setter to test
-     * @param array{0: string, 1: string} $getters       the getters to test
-     * @param float|int|string            $actual        the actual value to set
-     * @param float|int                   $expected      the expected value
-     *
-     * @throws InvalidValueException it should NOT happen
-     */
-    #[DataProvider('easyValuesProvider')]
-    public function testSetters(string $abstractPoint, string $setter, array $getters, float|int|string $actual, float|int $expected): void
-    {
-        $firstGetter = $getters[0];
-        $secondGetter = $getters[1];
-
-        $point = new $abstractPoint(10, 10);
-        self::assertObjectHasMethod($point, $setter);
-        self::assertObjectHasMethod($point, $firstGetter);
-        self::assertObjectHasMethod($point, $secondGetter);
-
-        $point->{$setter}($actual);
-        static::assertSame($expected, $point->{$firstGetter}());
-        static::assertSame($expected, $point->{$secondGetter}());
-    }
-
-    /**
-     * @return \Generator<string, array{0: float|int|string}, null, void>
-     */
-    public static function easyValuesProvider(): \Generator
-    {
-        $points = [
-            'GeometricPoint' => GeometricPoint::class,
-            'GeographicPoint' => GeographicPoint::class,
-        ];
-
-        $methods = [
-            'setX' => ['getX', 'getLongitude'],
-            'setY' => ['getY', 'getLatitude'],
-            'setLongitude' => ['getX', 'getLongitude'],
-            'setLatitude' => ['getY', 'getLatitude'],
-        ];
-
-        $values = [
-            'int(20)' => ['actual' => 20, 'expected' => 20],
-            'float(20.0)' => ['actual' => 20.0, 'expected' => 20.0],
-            'string(20)' => ['actual' => '20', 'expected' => 20],
-            'string(20.0)' => ['actual' => '20.0', 'expected' => 20.0],
-        ];
-
-        foreach ($points as $className => $class) {
-            foreach ($methods as $setter => $getters) {
-                foreach ($values as $valueName => $value) {
-                    yield sprintf('%s with %s and %s', $className, $setter, $valueName) => [$class, $setter, $getters, $value['actual'], $value['expected']];
-                }
-            }
-        }
-    }
-
-    /**
-     * This test checks that an exception is thrown when passing two coordinates separated by a space.
-     * This test checks all setters.
-     *
-     * @param class-string<AbstractPoint> $abstractPoint Geometric point and geographic point
-     * @param string                      $method        the method to test
-     */
-    #[DataProvider('setterProvider')]
-    public function testSettersWithAnArray(string $abstractPoint, string $method): void
-    {
-        $point = new $abstractPoint(10, 10);
-
-        self::expectException(InvalidValueException::class);
-        self::expectExceptionMessage('Invalid coordinate value, coordinate cannot be an array.');
-        static::assertTrue(method_exists($point, $method), sprintf('Method "%s":"%s" does not exist.', $abstractPoint, $method));
-        $point->{$method}('10 20');
-    }
-
-    /**
-     * @return \Generator<string, array{0: class-string<AbstractPoint>, 1: string}, null, void>
-     */
-    public static function setterProvider(): \Generator
-    {
-        $points = [
-            'GeometricPoint' => GeometricPoint::class,
-            'GeographicPoint' => GeographicPoint::class,
-        ];
-        $methods = [
-            'setX',
-            'setY',
-            'setLongitude',
-            'setLatitude',
-        ];
-        foreach ($points as $className => $class) {
-            foreach ($methods as $method) {
-                yield sprintf('%s with %s', $className, $method) => [$class, $method];
-            }
-        }
-    }
-
-    /**
-     * Test to convert point to array.
-     *
-     * @param class-string<AbstractPoint> $abstractPoint Geometric point and geographic point
-     */
-    #[DataProvider('pointTypeProvider')]
-    public function testToArray(string $abstractPoint): void
-    {
-        $expected = [-10, 11];
-        $point = new $abstractPoint(-10, 11);
-        $actual = $point->toArray();
-        static::assertEquals($expected, $actual);
-
-        $expected = [-42.42, 42.43];
-        $point = new $abstractPoint(-42.42, 42.43);
-        $actual = $point->toArray();
-        static::assertEquals($expected, $actual);
-    }
-
-    /**
-     * Test bad string parameters - Two invalid parameters.
-     *
-     * @param class-string<AbstractPoint> $abstractPoint Geometric point and geographic point
-     */
-    #[DataProvider('pointTypeProvider')]
-    public function testTwoInvalidArguments(string $abstractPoint): void
-    {
-        $this->expectException(InvalidValueException::class);
-        $this->expectExceptionMessage(sprintf('Invalid parameters passed to %s::__construct: NULL, NULL', $abstractPoint));
-
-        new $abstractPoint(null, null);
-    }
-
-    /**
-     * Test bad string parameters - More than 3 parameters.
-     *
-     * @param class-string<AbstractPoint> $abstractPoint Geometric point and geographic point
-     */
-    #[DataProvider('pointTypeProvider')]
-    public function testUnusedArguments(string $abstractPoint): void
-    {
-        $this->expectException(InvalidValueException::class);
-        $this->expectExceptionMessage(sprintf('Invalid parameters passed to %s::__construct: 1, 2, 3, 4, NULL, 5', $abstractPoint));
-
-        new $abstractPoint(1, 2, 3, 4, null, 5);
-    }
-
-    /**
-     * @return \Generator<string, array{0: class-string<AbstractPoint>}, null, void>
-     */
-    public static function pointTypeProvider(): \Generator
-    {
-        yield 'GeometricPoint' => [GeometricPoint::class];
-
-        yield 'GeographicPoint' => [GeographicPoint::class];
-    }
-}
