+++ tests/LongitudeOne/Spatial/Tests/PHP/Types/AbstractPointTestCase.php
@@ -1,0 +1,55 @@
+<?php
+/**
+ * This file is part of the Doctrine Spatial extension.
+ *
+ * PHP 8.4 | 8.5
+ * Doctrine ORM ^3.6
+ *
+ * Copyright Alexandre Tranchant <alexandre.tranchant@gmail.com> 2017-2026
+ * Copyright Longitude One 2020-2026
+ * Copyright 2015 Derek J. Lambert
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ */
+
+declare(strict_types=1);
+
+namespace LongitudeOne\Spatial\Tests\PHP\Types;
+
+use LongitudeOne\Spatial\PHP\Types\AbstractPoint;
+use LongitudeOne\Spatial\PHP\Types\Geography\Point as GeographicPoint;
+use LongitudeOne\Spatial\PHP\Types\Geometry\Point as GeometricPoint;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * Shared subject configuration for AbstractPoint contract tests.
+ *
+ * @internal
+ */
+abstract class AbstractPointTestCase extends TestCase
+{
+    protected const DELTA = 0.000000000001;
+
+    /**
+     * @return \Generator<string, array{0: class-string<AbstractPoint>}, null, void>
+     */
+    public static function pointTypeProvider(): \Generator
+    {
+        yield 'GeometricPoint' => [GeometricPoint::class];
+
+        yield 'GeographicPoint' => [GeographicPoint::class];
+    }
+
+    /**
+     * Assert that the object has the method.
+     *
+     * @param object $object object to test
+     * @param string $method the method to test
+     */
+    protected static function assertObjectHasMethod(object $object, string $method): void
+    {
+        static::assertTrue(method_exists($object, $method), sprintf('Method "%s":"%s" does not exist.', $object::class, $method));
+    }
+}
